app/vmui/logs: fix zoom behavior on VictoriaLogs chart
