+++ app/vmui/packages/vmui/src/components/Chart/BarHitsChart/BarHitsChart.tsx
@@ -1,23 +1,13 @@
-import React, { FC, useCallback, useMemo, useRef, useState } from "preact/compat";
+import React, { FC, useState } from "preact/compat";
 import "./style.scss";
 import "uplot/dist/uPlot.min.css";
-import useElementSize from "../../../hooks/useElementSize";
-import uPlot, { AlignedData } from "uplot";
-import { useEffect } from "react";
-import useBarHitsOptions, { getLabelFromLogHit } from "./hooks/useBarHitsOptions";
-import BarHitsTooltip from "./BarHitsTooltip/BarHitsTooltip";
+import { AlignedData } from "uplot";
 import { TimeParams } from "../../../types";
-import usePlotScale from "../../../hooks/uplot/usePlotScale";
-import useReadyChart from "../../../hooks/uplot/useReadyChart";
-import useZoomChart from "../../../hooks/uplot/useZoomChart";
 import classNames from "classnames";
-import { LegendLogHits, LogHits } from "../../../api/types";
-import { addSeries, delSeries, setBand } from "../../../utils/uplot";
+import { LogHits } from "../../../api/types";
 import { GraphOptions, GRAPH_STYLES } from "./types";
 import BarHitsOptions from "./BarHitsOptions/BarHitsOptions";
-import stack from "../../../utils/uplot/stack";
-import BarHitsLegend from "./BarHitsLegend/BarHitsLegend";
-import { calculateTotalHits, sortLogHits } from "../../../utils/logs";
+import BarHitsPlot from "./BarHitsPlot/BarHitsPlot";
 
 interface Props {
   logHits: LogHits[];
@@ -27,9 +17,6 @@ interface Props {
   onApplyFilter: (value: string) => void;
 }
 const BarHitsChart: FC<Props> = ({ logHits, data: _data, period, setPeriod, onApplyFilter }) => {
-  const [containerRef, containerSize] = useElementSize();
-  const uPlotRef = useRef<HTMLDivElement>(null);
-  const [uPlotInst, setUPlotInst] = useState<uPlot>();
   const [graphOptions, setGraphOptions] = useState<GraphOptions>({
     graphStyle: GRAPH_STYLES.LINE_STEPPED,
     stacked: false,
@@ -37,100 +24,6 @@ const BarHitsChart: FC<Props> = ({ logHits, data: _data, period, setPeriod, onAp
     hideChart: false,
   });
 
-  const { xRange, setPlotScale } = usePlotScale({ period, setPeriod });
-  const { onReadyChart, isPanning } = useReadyChart(setPlotScale);
-  useZoomChart({ uPlotInst, xRange, setPlotScale });
-
-  const isEmptyData = useMemo(() => _data.every(d => d.length === 0), [_data]);
-
-  const { data, bands } = useMemo(() => {
-    return graphOptions.stacked ? stack(_data, () => false) : { data: _data, bands: [] };
-  }, [graphOptions, _data]);
-
-  const { options, series, focusDataIdx } = useBarHitsOptions({
-    data,
-    logHits,
-    bands,
-    xRange,
-    containerSize,
-    onReadyChart,
-    setPlotScale,
-    graphOptions
-  });
-
-  const prepareLegend = useCallback((hits: LogHits[], totalHits: number): LegendLogHits[] => {
-    return hits.map((hit) => {
-      const label = getLabelFromLogHit(hit);
-
-      const legendItem: LegendLogHits = {
-        label,
-        isOther: hit._isOther,
-        fields: hit.fields,
-        total: hit.total || 0,
-        totalHits,
-        stroke: series.find((s) => s.label === label)?.stroke,
-      };
-
-      return legendItem;
-    }).sort(sortLogHits("total"));
-  }, [series]);
-
-
-  const legendDetails: LegendLogHits[] = useMemo(() => {
-    const totalHits = calculateTotalHits(logHits);
-    return prepareLegend(logHits, totalHits);
-  }, [logHits, prepareLegend]);
-
-  useEffect(() => {
-    if (!uPlotInst) return;
-
-    const oldSeriesMap = new Map(uPlotInst.series.map(s => [s.label, s]));
-
-    const syncedSeries = series.map(s => {
-      const old = oldSeriesMap.get(s.label);
-      return old ? { ...s, show: old.show } : s;
-    });
-
-    delSeries(uPlotInst);
-    addSeries(uPlotInst, syncedSeries, true);
-    setBand(uPlotInst, syncedSeries);
-    uPlotInst.redraw();
-  }, [series, uPlotInst]);
-
-
-  useEffect(() => {
-    if (!uPlotInst) return;
-    uPlotInst.delBand();
-    bands.forEach(band => {
-      uPlotInst.addBand(band);
-    });
-    uPlotInst.redraw();
-  }, [bands]);
-
-  useEffect(() => {
-    if (!uPlotRef.current) return;
-    const uplot = new uPlot(options, data, uPlotRef.current);
-    setUPlotInst(uplot);
-    return () => uplot.destroy();
-  }, [uPlotRef.current]);
-
-  useEffect(() => {
-    if (!uPlotInst) return;
-    uPlotInst.scales.x.range = () => [xRange.min, xRange.max];
-    uPlotInst.redraw();
-  }, [xRange]);
-
-  useEffect(() => {
-    if (!uPlotInst) return;
-    uPlotInst.setSize(containerSize);
-    uPlotInst.redraw();
-  }, [containerSize]);
-
-  useEffect(() => {
-    if (!uPlotInst) return;
-    uPlotInst.setData(data);
-    uPlotInst.redraw();
-  }, [data]);
 
   return (
     <div
@@ -140,32 +33,16 @@ const BarHitsChart: FC<Props> = ({ logHits, data: _data, period, setPeriod, onAp
       })}
     >
       {!graphOptions.hideChart && (
-        <div
-          className={classNames({
-            "vm-bar-hits-chart": true,
-            "vm-bar-hits-chart_panning": isPanning
-          })}
-          ref={containerRef}
-        >
-          <div
-            className="vm-line-chart__u-plot"
-            ref={uPlotRef}
-          />
-          <BarHitsTooltip
-            uPlotInst={uPlotInst}
-            data={_data}
-            focusDataIdx={focusDataIdx}
-          />
-        </div>
-      )}
-      <BarHitsOptions onChange={setGraphOptions}/>
-      {uPlotInst && !isEmptyData && !graphOptions.hideChart && (
-        <BarHitsLegend
-          uPlotInst={uPlotInst}
+        <BarHitsPlot
+          logHits={logHits}
+          data={_data}
+          period={period}
+          setPeriod={setPeriod}
           onApplyFilter={onApplyFilter}
-          legendDetails={legendDetails}
+          graphOptions={graphOptions}
         />
       )}
+      <BarHitsOptions onChange={setGraphOptions}/>
     </div>
   );
 };
